move to o.e.sh.binding.mqtt.generic folder
