+++ extensions/binding/org.eclipse.smarthome.binding.mqtt.generic.test/META-INF/MANIFEST.MF
@@ -1,0 +1,26 @@
+Manifest-Version: 1.0
+Bundle-ManifestVersion: 2
+Bundle-Name: MQTT Binding Tests
+Bundle-SymbolicName: org.eclipse.smarthome.binding.mqtt.generic.test;singleton:=true
+Bundle-Vendor: Eclipse.org/SmartHome
+Bundle-Version: 0.10.0.qualifier
+Bundle-RequiredExecutionEnvironment: JavaSE-1.8
+Fragment-Host: org.eclipse.smarthome.binding.mqtt.generic
+Import-Package: com.jayway.jsonpath,
+ net.minidev.json,
+ net.minidev.json.parser,
+ org.eclipse.smarthome.binding.mqtt.generic,
+ org.eclipse.smarthome.binding.mqtt.generic.handler,
+ org.eclipse.smarthome.core.common.registry,
+ org.eclipse.smarthome.core.events,
+ org.eclipse.smarthome.core.thing,
+ org.eclipse.smarthome.core.thing.binding.builder,
+ org.eclipse.smarthome.core.thing.util,
+ org.eclipse.smarthome.core.types,
+ org.hamcrest;core=split,
+ org.junit;version="4.0.0",
+ org.mockito,
+ org.mockito.stubbing,
+ org.osgi.framework,
+ org.osgi.service.device,
+ org.slf4j
