+++ extensions/binding/org.eclipse.smarthome.binding.mqtt.generic.test/build.properties
