+++ extensions/binding/org.eclipse.smarthome.binding.mqtt.generic/.classpath
