+++ extensions/binding/org.eclipse.smarthome.binding.mqtt.generic/ESH-INF/binding/binding.xml
