+++ extensions/binding/org.eclipse.smarthome.binding.mqtt.generic/ESH-INF/i18n/mqttbroker_de.properties
