+++ extensions/binding/org.eclipse.smarthome.binding.mqtt.generic/ESH-INF/thing/thing-types.xml
