+++ extensions/binding/org.eclipse.smarthome.binding.mqtt.generic/OSGI-INF/.gitignore
