+++ extensions/binding/org.eclipse.smarthome.binding.mqtt.generic/README.md
