+++ extensions/binding/org.eclipse.smarthome.binding.mqtt.generic/build.properties
