+++ features/karaf/esh-ext/src/main/feature/feature.xml
@@ -67,7 +67,7 @@
   <feature name="esh-binding-mqtt" version="${project.version}">
     <feature>esh-base</feature>
     <feature>esh-io-transport-mqtt</feature>
-    <bundle>mvn:org.eclipse.smarthome.binding/org.eclipse.smarthome.binding.mqtt/${project.version}</bundle>
+    <bundle>mvn:org.eclipse.smarthome.binding/org.eclipse.smarthome.binding.mqtt.generic/${project.version}</bundle>
   </feature>
 
   <feature name="esh-binding-ntp" version="${project.version}">
