+++ features/org.eclipse.smarthome.feature.runtime.binding/feature.xml
@@ -81,7 +81,7 @@
          unpack="false"/>
 
    <plugin
-         id="org.eclipse.smarthome.binding.mqtt"
+         id="org.eclipse.smarthome.binding.mqtt.generic"
          download-size="0"
          install-size="0"
          version="0.0.0"
